# migration-package/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAISON NOIR Parfumerie API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Perfume {
  id: number;
  name: string;
  brand: string;
  price_per_ml: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  image_url?: string | null;
  /** @nullable */
  category_id: number | null;
  /** @nullable */
  category_name?: string | null;
  in_stock: boolean;
  is_featured?: boolean;
  created_at: string;
}

export interface PerfumeInput {
  name: string;
  brand: string;
  price_per_ml: number;
  description?: string;
  notes?: string;
  image_url?: string;
  /** @nullable */
  category_id?: number | null;
  in_stock?: boolean;
  is_featured?: boolean;
}

export interface PerfumeUpdate {
  name?: string;
  brand?: string;
  price_per_ml?: number;
  description?: string;
  notes?: string;
  image_url?: string;
  /** @nullable */
  category_id?: number | null;
  in_stock?: boolean;
  is_featured?: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image_url?: string | null;
  sort_order?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  image_url?: string;
  sort_order?: number;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
  image_url?: string;
  sort_order?: number;
}

export interface OrderItem {
  perfume_id: number;
  perfume_name: string;
  brand: string;
  price_per_ml: number;
  ml: number;
  subtotal: number;
}

export interface Order {
  id: number;
  customer_name: string;
  customer_phone: string;
  /** @nullable */
  customer_address?: string | null;
  /** @nullable */
  customer_comment?: string | null;
  status: string;
  total: number;
  items: OrderItem[];
  created_at: string;
}

export interface OrderItemInput {
  perfume_id: number;
  ml: number;
}

export interface OrderInput {
  customer_name: string;
  customer_phone: string;
  customer_address?: string;
  customer_comment?: string;
  items: OrderItemInput[];
}

export interface OrderUpdate {
  status?: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  role?: string | null;
}

export interface Stats {
  total_orders: number;
  pending_orders: number;
  total_perfumes: number;
  total_revenue: number;
  recent_orders: Order[];
}

export interface DeleteResult {
  success: boolean;
}

export type ListPerfumesParams = {
/**
 * @nullable
 */
category_id?: number | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
sort?: string | null;
};

export type ListOrdersParams = {
/**
 * @nullable
 */
status?: string | null;
};