# migration-package/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAISON NOIR Parfumerie API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /perfumes:
    get:
      operationId: listPerfumes
      tags: [perfumes]
      summary: List all perfumes
      parameters:
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of perfumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Perfume"
    post:
      operationId: createPerfume
      tags: [perfumes]
      summary: Create a new perfume (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerfumeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Perfume"
  /perfumes/{id}:
    get:
      operationId: getPerfume
      tags: [perfumes]
      summary: Get a perfume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Perfume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Perfume"
        "404":
          description: Not found
    patch:
      operationId: updatePerfume
      tags: [perfumes]
      summary: Update a perfume (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerfumeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Perfume"
    delete:
      operationId: deletePerfume
      tags: [perfumes]
      summary: Delete a perfume (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
  schemas:
...
    Perfume:
      type: object
      required: [id, name, brand, price_per_ml, category_id, in_stock, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price_per_ml:
          type: number
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        image_url:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        category_name:
          type: ["string", "null"]
        in_stock:
          type: boolean
        is_featured:
          type: boolean
        created_at:
          type: string
    PerfumeInput:
      type: object
      required: [name, brand, price_per_ml]
      properties:
        name:
          type: string
        brand:
          type: string
        price_per_ml:
          type: number
        description:
          type: string
        notes:
          type: string
        image_url:
          type: string
        category_id:
          type: ["integer", "null"]
        in_stock:
          type: boolean
        is_featured:
          type: boolean
    PerfumeUpdate:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        price_per_ml:
          type: number
        description:
          type: string
        notes:
          type: string
        image_url:
          type: string
        category_id:
          type: ["integer", "null"]
        in_stock:
          type: boolean
        is_featured:
          type: boolean
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean