# migration-package/lib/api-zod/src/generated/types/perfumeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MAISON NOIR Parfumerie API
 * OpenAPI spec version: 0.1.0
 */

export interface PerfumeUpdate {
  name?: string;
  brand?: string;
  price_per_ml?: number;
  description?: string;
  notes?: string;
  image_url?: string;
  /** @nullable */
  category_id?: number | null;
  in_stock?: boolean;
  is_featured?: boolean;
}